requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
Django==5.1.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==24.3
django-nine==0.2.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
packaging==24.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
